# docs/swagger.yaml
basePath: /api/
definitions:
  reqs.OrderPostBuy:
    properties:
      price:
        default: 10
        type: number
      quantity:
        default: 5
        type: integer
    required:
    - price
    - quantity
    type: object
  reqs.OrderPostSell:
    properties:
      price:
        default: 10
        type: number
      quantity:
        default: 5
        type: integer
    required:
    - price
    - quantity
    type: object
  reqs.OrderPostTest:
    properties:
      price:
        default: 10
        type: number
      quantity:
        default: 5
        type: integer
      run_times:
        default: 10000
        type: integer
    required:
    - price
    - quantity
    - run_times
    type: object
  resp.Base:
    properties:
      data: {}
      event_id:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: homework
  title: zerologix-homework
  version: "1.0"
paths:
  /trade/order/buy:
    post:
      consumes:
      - application/json
      description: 買
      parameters:
      - description: 參數
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/reqs.OrderPostBuy'
      produces:
      - application/json
      responses:
        "200":
          description: 資料
          schema:
            $ref: '#/definitions/resp.Base'
      security:
      - ApiKeyAuth: []
      summary: 買
      tags:
      - Order
  /trade/order/sell:
    post:
      consumes:
      - application/json
      description: 賣
      parameters:
      - description: 參數
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/reqs.OrderPostSell'
      produces:
      - application/json
      responses:
        "200":
          description: 資料
          schema:
            $ref: '#/definitions/resp.Base'
      security:
      - ApiKeyAuth: []
      summary: 賣
      tags:
      - Order
  /trade/order/test:
    post:
      consumes:
      - application/json
      description: 測試
      parameters:
      - description: 參數
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/reqs.OrderPostTest'
      produces:
      - application/json
      responses:
        "200":
          description: 資料
          schema:
            $ref: '#/definitions/resp.Base'
      security:
      - ApiKeyAuth: []
      summary: 測試
      tags:
      - Order
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"